fix: Restore persistent error squiggles after later build steps run
The fix for corrupt/incomplete output files at commit c5d21f6 (in
version 0.2.1) seems to have broken error squiggles. The behavior after
that change is that an error will briefly be displayed, but since later
build steps run unconditionally (by design), VSCode assumes it's OK to
clear the error squiggles.

It turns out that this has to do with the Problem Matcher and/or
Language ID being reused. Since all the CPSA tools use the same error
format, we reused the Problem Matcher. But that implicitly told VSCode
that we were "re-running" a build step and the previous errors were no
longer needed.

Experimentally, I determined that one must use separate Problem
Matchers, and also separate Language IDs in the "owner" fields, in order
to convince VSCode to treat the errors as separate groups. Therefore,
this commit adds three new Language IDs and three copies of the same
Problem Matcher. It looks silly but it works.